move helperCreateLiveKitRoom into helper.go
