+++ helper.go
@@ -40,6 +40,37 @@ func NewUniqueID() UniqueID {
 	return UniqueID(base32.HexEncoding.WithPadding(base32.NoPadding).EncodeToString(b))
 }
 
+var helperCreateLiveKitRoom = func(ctx context.Context, liveKitAuth *LiveKitAuth, room LiveKitRoomAlias, matrixUser string, lkIdentity LiveKitIdentity) error {
+	roomClient := lksdk.NewRoomServiceClient(liveKitAuth.lkUrl, liveKitAuth.key, liveKitAuth.secret)
+	creationStart := time.Now().Unix()
+	lkRoom, err := roomClient.CreateRoom(
+		ctx,
+		&livekit.CreateRoomRequest{
+			Name:             string(room),
+			EmptyTimeout:     5 * 60, // 5 Minutes to keep the room open if no one joins
+			DepartureTimeout: 20,     // number of seconds to keep the room after everyone leaves
+			MaxParticipants:  0,      // 0 == no limitation
+		},
+	)
+
+	if err != nil {
+		return fmt.Errorf("unable to create room %s: %w", room, err)
+	}
+
+	// Log the room creation time and the user info
+	isNewRoom := lkRoom.GetCreationTime() >= creationStart && lkRoom.GetCreationTime() <= time.Now().Unix()
+	slog.Info(
+		fmt.Sprintf("createLiveKitRoom: %s Room", map[bool]string{true: "Created", false: "Using"}[isNewRoom]),
+		"room", room,
+		"roomSid", lkRoom.Sid,
+		"lkId", lkIdentity,
+		"matrixUser", matrixUser,
+		"access", "full",
+	)
+
+	return nil
+}
+
 var helperLiveKitParticipantLookup = func(ctx context.Context, lkAuth LiveKitAuth, lkRoomAlias LiveKitRoomAlias, lkId LiveKitIdentity, ch chan SFUMessage) (bool, error) {
 	roomClient := lksdk.NewRoomServiceClient(
 		lkAuth.lkUrl,
